Archivos modificados

+++ tp-persona/src/country/country.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Country } from '../entities/country';
@@ -18,21 +18,33 @@ export class CountryService {
     );
   }
 
-  findOne(id: number) {
-    return this.repo.findOneByOrFail({ id });
+  async findOne(id: number) {
+    const country = await this.repo.findOneBy({ id });
+    if (!country) return null;
+    return {
+      id: country.id,
+      name: country.name,
+    };
   }
 
-  create(data: Partial<Country>) {
-    return this.repo.save(this.repo.create(data));
+  async create(data: Partial<Country>) {
+    const country = await this.repo.save(this.repo.create(data));
+    return this.findOne(country.id);
   }
 
   async update(id: number, data: Partial<Country>) {
-    const entity = await this.findOne(id);
-    return this.repo.save({ ...entity, ...data });
+    const entity = await this.repo.findOneBy({ id });
+    if (!entity) return null;
+    await this.repo.save({ ...entity, ...data });
+    return this.findOne(id);
   }
 
   async remove(id: number) {
-    const entity = await this.findOne(id);
-    return this.repo.remove(entity);
+    const entity = await this.repo.findOneBy({ id });
+    if (entity) {
+      await this.repo.remove(entity);
+      return { message: 'deleted' };
+    }
+    return null;
   }
 }
